scraping/requirements.txt
requests==2.24.0
beautifulsoup4==4.9.1
psycopg2-binary